Harden block-runtime metadata core boundaries
