+++ packages/wp-typia-block-runtime/src/metadata-core.ts
@@ -3,7 +3,23 @@ import type {} from './typia-tags.js';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 
-import { getTaggedSyncBlockMetadataFailureCode } from './metadata-analysis.js';
+import {
+  type GeneratedArtifactFile,
+  type GeneratedArtifactDriftIssue,
+  normalizeSyncBlockMetadataFailure,
+  reconcileGeneratedArtifacts,
+  resolveSyncBlockMetadataPaths,
+} from './metadata-core-artifacts.js';
+import {
+  assertValidClientIdentifier,
+  normalizeSyncEndpointClientOptions,
+  normalizeSyncRestOpenApiOptions,
+  reserveUniqueClientTypeIdentifier,
+  resolveEndpointClientContract,
+  toJavaScriptStringLiteral,
+  toModuleImportPath,
+  toValidatorAccessExpression,
+} from './metadata-core-endpoint-client.js';
 import { renderPhpValidator } from './metadata-php-render.js';
 import { analyzeSourceType, analyzeSourceTypes } from './metadata-parser.js';
 import {
@@ -275,13 +291,6 @@ export interface SyncRestOpenApiResult {
   schemaNames: string[];
 }
 
-interface NormalizedSyncRestOpenApiOptions {
-  manifest: EndpointManifestDefinition;
-  openApiPath: string;
-  projectRoot: string;
-  typesFile: string;
-}
-
 interface SyncEndpointClientBaseOptions {
   /** Output path for the generated portable client module. */
   clientFile: string;
@@ -311,122 +320,6 @@ export interface SyncEndpointClientResult {
   clientPath: string;
   /** Operation ids emitted as endpoint constants and convenience wrappers. */
   operationIds: string[];
-}
-
-interface NormalizedSyncEndpointClientOptions {
-  clientPath: string;
-  manifest: EndpointManifestDefinition;
-  projectRoot: string;
-  typesFile: string;
-  validatorsFile: string;
-}
-
-interface ResolvedSyncBlockMetadataPaths {
-  blockJsonPath: string;
-  jsonSchemaPath: string | null;
-  manifestPath: string;
-  openApiPath: string | null;
-  phpValidatorPath: string;
-  projectRoot: string;
-}
-
-interface GeneratedArtifactFile {
-  content: string;
-  path: string;
-}
-
-interface GeneratedArtifactDriftIssue {
-  path: string;
-  reason: 'missing' | 'stale';
-}
-
-class GeneratedArtifactDriftError extends Error {
-  readonly issues: GeneratedArtifactDriftIssue[];
-
-  constructor(issues: GeneratedArtifactDriftIssue[]) {
-    const detail = issues
-      .map((issue) => `- ${issue.path} (${issue.reason})`)
-      .join('\n');
-
-    super(`Generated artifacts are missing or stale:\n${detail}`);
-    this.name = 'GeneratedArtifactDriftError';
-    this.issues = issues;
-  }
-}
-
-function reconcileGeneratedArtifacts(
-  artifacts: readonly GeneratedArtifactFile[],
-  executionOptions: ArtifactSyncExecutionOptions | undefined,
-  preexistingIssues: GeneratedArtifactDriftIssue[] = [],
-): void {
-  if (executionOptions?.check !== true) {
-    for (const artifact of artifacts) {
-      fs.mkdirSync(path.dirname(artifact.path), { recursive: true });
-      fs.writeFileSync(artifact.path, artifact.content, 'utf8');
-    }
-
-    return;
-  }
-
-  const issues = [...preexistingIssues];
-
-  for (const artifact of artifacts) {
-    if (!fs.existsSync(artifact.path)) {
-      issues.push({
-        path: artifact.path,
-        reason: 'missing',
-      });
-      continue;
-    }
-
-    const currentContent = fs.readFileSync(artifact.path, 'utf8');
-    if (
-      normalizeGeneratedArtifactContentForComparison(currentContent) !==
-      normalizeGeneratedArtifactContentForComparison(artifact.content)
-    ) {
-      issues.push({
-        path: artifact.path,
-        reason: 'stale',
-      });
-    }
-  }
-
-  if (issues.length > 0) {
-    throw new GeneratedArtifactDriftError(issues);
-  }
-}
-
-function normalizeGeneratedArtifactContentForComparison(content: string): string {
-  return content.replace(/\r\n?/g, '\n');
-}
-
-function resolveSyncBlockMetadataPaths(
-  options: SyncBlockMetadataOptions,
-): ResolvedSyncBlockMetadataPaths {
-  const projectRoot = path.resolve(options.projectRoot ?? process.cwd());
-  const blockJsonPath = path.resolve(projectRoot, options.blockJsonFile);
-  const manifestRelativePath =
-    options.manifestFile ??
-    path.join(path.dirname(options.blockJsonFile), 'typia.manifest.json');
-  const manifestPath = path.resolve(projectRoot, manifestRelativePath);
-  const phpValidatorPath = path.resolve(
-    projectRoot,
-    options.phpValidatorFile ??
-      path.join(path.dirname(manifestRelativePath), 'typia-validator.php'),
-  );
-
-  return {
-    blockJsonPath,
-    jsonSchemaPath: options.jsonSchemaFile
-      ? path.resolve(projectRoot, options.jsonSchemaFile)
-      : null,
-    manifestPath,
-    openApiPath: options.openApiFile
-      ? path.resolve(projectRoot, options.openApiFile)
-      : null,
-    phpValidatorPath,
-    projectRoot,
-  };
 }
 
 /**
@@ -1086,305 +979,4 @@ export async function syncEndpointClient(
     endpointCount: manifest.endpoints.length,
     operationIds: [...operationIds],
   };
-}
-
-function normalizeSyncRestOpenApiOptions(
-  options: SyncRestOpenApiOptions,
-): NormalizedSyncRestOpenApiOptions {
-  const projectRoot = path.resolve(options.projectRoot ?? process.cwd());
-  const openApiPath = path.resolve(projectRoot, options.openApiFile);
-
-  if ('manifest' in options) {
-    const hasDecomposedInputs =
-      'contracts' in options ||
-      'endpoints' in options ||
-      'openApiInfo' in options;
-    if (hasDecomposedInputs) {
-      throw new Error(
-        'syncRestOpenApi() accepts either { manifest, ... } or { contracts, endpoints, ... }, but not both.',
-      );
-    }
-    if (options.manifest == null) {
-      throw new Error(
-        'syncRestOpenApi() requires a manifest object when using { manifest, ... }.',
-      );
-    }
-
-    return {
-      manifest: options.manifest,
-      openApiPath,
-      projectRoot,
-      typesFile: options.typesFile,
-    };
-  }
-
-  return {
-    manifest: {
-      contracts: options.contracts,
-      endpoints: options.endpoints,
-      ...(options.openApiInfo ? { info: options.openApiInfo } : {}),
-    },
-    openApiPath,
-    projectRoot,
-    typesFile: options.typesFile,
-  };
-}
-
-function normalizeSyncEndpointClientOptions(
-  options: SyncEndpointClientOptions,
-): NormalizedSyncEndpointClientOptions {
-  const projectRoot = path.resolve(options.projectRoot ?? process.cwd());
-  const clientPath = path.resolve(projectRoot, options.clientFile);
-  const typesFile = path.resolve(projectRoot, options.typesFile);
-  const inferredValidatorsFile =
-    options.validatorsFile ??
-    (() => {
-      const nextPath = options.typesFile.replace(/api-types\.ts$/u, 'api-validators.ts');
-      if (nextPath === options.typesFile) {
-        throw new Error(
-          'syncEndpointClient() could not infer validatorsFile from typesFile; pass validatorsFile explicitly.',
-        );
-      }
-
-      return nextPath;
-    })();
-  const validatorsFile = path.resolve(projectRoot, inferredValidatorsFile);
-
-  if (!fs.existsSync(typesFile)) {
-    throw new Error(`Unable to generate an endpoint client because the types file does not exist: ${typesFile}`);
-  }
-  if (!fs.existsSync(validatorsFile)) {
-    throw new Error(
-      `Unable to generate an endpoint client because the validators file does not exist: ${validatorsFile}`,
-    );
-  }
-
-  return {
-    clientPath,
-    manifest: options.manifest,
-    projectRoot,
-    typesFile: options.typesFile,
-    validatorsFile: path.relative(projectRoot, validatorsFile),
-  };
-}
-
-function resolveEndpointClientContract(
-  manifest: EndpointManifestDefinition,
-  contractKey: string,
-  operationId: string,
-  fieldName: 'bodyContract' | 'queryContract' | 'responseContract',
-): EndpointManifestContractDefinition {
-  const contract = manifest.contracts[contractKey];
-  if (!contract) {
-    throw new Error(
-      `Endpoint "${operationId}" references missing ${fieldName} "${contractKey}" while generating the endpoint client.`,
-    );
-  }
-
-  return contract;
-}
-
-function toClientPropertyName(value: string): string {
-  return value
-    .replace(/[^A-Za-z0-9]+(.)/g, (_match, next: string) => next.toUpperCase())
-    .replace(/^[A-Z]/, (match) => match.toLowerCase());
-}
-
-function toValidatorAccessExpression(
-  contractKey: string,
-  seenPropertyNames: Map<string, string>,
-): string {
-  const propertyName = toClientPropertyName(contractKey);
-  const previousContractKey = seenPropertyNames.get(propertyName);
-
-  if (previousContractKey && previousContractKey !== contractKey) {
-    throw new Error(
-      `Contract keys "${previousContractKey}" and "${contractKey}" both normalize to apiValidators[${toJavaScriptStringLiteral(
-        propertyName,
-      )}] while generating the endpoint client.`,
-    );
-  }
-
-  seenPropertyNames.set(propertyName, contractKey);
-  return /^[$A-Z_][0-9A-Z_$]*$/iu.test(propertyName)
-    ? `apiValidators.${propertyName}`
-    : `apiValidators[${toJavaScriptStringLiteral(propertyName)}]`;
-}
-
-function toModuleImportPath(fromFile: string, targetFile: string): string {
-  let relativePath = path.relative(path.dirname(fromFile), targetFile).replace(/\\/g, '/');
-  if (!relativePath.startsWith('.')) {
-    relativePath = `./${relativePath}`;
-  }
-
-  return relativePath.replace(/\.[^.]+$/u, '');
-}
-
-function toJavaScriptStringLiteral(value: string): string {
-  return `'${value
-    .replace(/\\/g, '\\\\')
-    .replace(/'/g, "\\'")
-    .replace(/\r/g, '\\r')
-    .replace(/\n/g, '\\n')}'`;
-}
-
-const RESERVED_CLIENT_IDENTIFIERS = new Set([
-  'await',
-  'break',
-  'case',
-  'catch',
-  'class',
-  'const',
-  'continue',
-  'debugger',
-  'default',
-  'delete',
-  'do',
-  'else',
-  'enum',
-  'export',
-  'extends',
-  'false',
-  'finally',
-  'for',
-  'function',
-  'if',
-  'implements',
-  'import',
-  'in',
-  'interface',
-  'instanceof',
-  'let',
-  'new',
-  'null',
-  'package',
-  'private',
-  'protected',
-  'public',
-  'return',
-  'static',
-  'super',
-  'switch',
-  'this',
-  'throw',
-  'true',
-  'try',
-  'typeof',
-  'var',
-  'void',
-  'while',
-  'with',
-  'yield',
-]);
-
-/**
- * Guard generated client identifiers so emitted modules remain valid JavaScript.
- *
- * @param value Candidate identifier to validate before code generation.
- * @param label Human-readable label for error reporting.
- * @throws {Error} When the identifier is syntactically invalid or reserved.
- */
-function assertValidClientIdentifier(value: string, label: string): void {
-  if (!/^[$A-Z_][0-9A-Z_$]*$/iu.test(value)) {
-    throw new Error(
-      `Generated endpoint client ${label} "${value}" is not a valid JavaScript identifier.`,
-    );
-  }
-  if (RESERVED_CLIENT_IDENTIFIERS.has(value)) {
-    throw new Error(
-      `Generated endpoint client ${label} "${value}" is a reserved JavaScript identifier.`,
-    );
-  }
-}
-
-function reserveUniqueClientTypeIdentifier(
-  preferred: string,
-  occupied: Set<string>,
-): string {
-  let suffix = 0;
-
-  while (true) {
-    const candidate =
-      suffix === 0
-        ? preferred
-        : suffix === 1
-          ? `${preferred}Alias`
-          : `${preferred}Alias${suffix}`;
-    if (!occupied.has(candidate) && !RESERVED_CLIENT_IDENTIFIERS.has(candidate)) {
-      assertValidClientIdentifier(candidate, 'type alias');
-      occupied.add(candidate);
-      return candidate;
-    }
-    suffix += 1;
-  }
-}
-
-function normalizeSyncBlockMetadataFailure(
-  error: unknown,
-): SyncBlockMetadataFailure {
-  if (error instanceof Error) {
-    return {
-      code: resolveSyncBlockMetadataFailureCode(error),
-      message: error.message,
-      name: error.name,
-    };
-  }
-
-  return {
-    code: 'unknown-internal-error',
-    message: String(error),
-    name: 'NonErrorThrow',
-  };
-}
-
-function resolveSyncBlockMetadataFailureCode(
-  error: Error,
-): SyncBlockMetadataFailureCode {
-  if (error instanceof GeneratedArtifactDriftError) {
-    return 'stale-generated-artifact';
-  }
-
-  const taggedCode = getTaggedSyncBlockMetadataFailureCode(error);
-  if (taggedCode) {
-    return taggedCode;
-  }
-
-  const { message } = error;
-  if (message.startsWith('Unsupported type node at ')) {
-    return 'unsupported-type-node';
-  }
-  if (
-    message.startsWith('Recursive types are not supported:') ||
-    message.startsWith('Recursive type')
-  ) {
-    return 'recursive-type';
-  }
-  if (
-    message.startsWith('Unable to load types file:') ||
-    message.startsWith('Unable to find source type "') ||
-    message.startsWith('Unable to resolve type reference "') ||
-    message.includes('must resolve to an object shape')
-  ) {
-    return 'invalid-source-type';
-  }
-  if (
-    message.startsWith('Unsupported ') ||
-    message.startsWith('Mixed primitive enums are not supported at ') ||
-    message.startsWith('Indexed access ') ||
-    message.startsWith('Intersection at ') ||
-    message.startsWith('WordPress extraction ') ||
-    message.startsWith('Generic type declarations are not supported at ') ||
-    message.startsWith('Generic type references are not supported at ') ||
-    message.startsWith('Class and enum references are not supported at ') ||
-    message.startsWith('Discriminated union at ') ||
-    message.startsWith('External or non-serializable ') ||
-    message.startsWith('Conflicting ') ||
-    message.startsWith('Tag "') ||
-    message.startsWith('Only object-like interface extensions are supported:') ||
-    message.startsWith('Array type is missing an item type at ')
-  ) {
-    return 'unsupported-type-pattern';
-  }
-
-  return 'unknown-internal-error';
 }
