+++ packages/wp-typia-block-runtime/tests/block-runtime.test.ts
@@ -53,14 +53,25 @@ describe("@wp-typia/block-runtime", () => {
 		expect(typeof rootModule.createAttributeUpdater).toBe("function");
 		expect(typeof rootModule.createNestedAttributeUpdater).toBe("function");
 		expect(typeof blocksModule.createTypiaWebpackConfig).toBe("function");
+		expect(typeof blocksModule.buildScaffoldBlockRegistration).toBe("function");
+		expect(typeof blocksModule.defineScaffoldBlockMetadata).toBe("function");
+		expect(typeof blocksModule.assertScaffoldBlockMetadata).toBe("function");
+		expect(typeof blocksModule.parseScaffoldBlockMetadata).toBe("function");
 		expect(typeof schemaCoreModule.manifestToJsonSchema).toBe("function");
 		expect(typeof schemaCoreModule.manifestToOpenApi).toBe("function");
+		expect(typeof schemaCoreModule.normalizeEndpointAuthDefinition).toBe("function");
 		expect(Object.keys(migrationTypesModule)).toEqual([]);
+		expect(typeof metadataCoreModule.defineEndpointManifest).toBe("function");
 		expect(typeof metadataCoreModule.runSyncBlockMetadata).toBe("function");
+		expect(typeof metadataCoreModule.syncEndpointClient).toBe("function");
 		expect(typeof metadataCoreModule.syncRestOpenApi).toBe("function");
 		expect(typeof defaultsModule.applyTemplateDefaultsFromManifest).toBe("function");
+		expect(typeof defaultsModule.defineManifestDefaultsDocument).toBe("function");
+		expect(typeof defaultsModule.assertManifestDefaultsDocument).toBe("function");
 		expect(typeof defaultsModule.parseManifestDefaultsDocument).toBe("function");
 		expect(typeof editorModule.createEditorModel).toBe("function");
+		expect(typeof editorModule.defineManifestDocument).toBe("function");
+		expect(typeof editorModule.assertManifestDocument).toBe("function");
 		expect(typeof editorModule.parseManifestDocument).toBe("function");
 		expect(typeof identifiersModule.generateBlockId).toBe("function");
 		expect(typeof identifiersModule.generatePublicWriteRequestId).toBe("function");
